+++ docs/common/conf.py
@@ -60,6 +60,7 @@ nitpick_ignore = [
     ('py:class', 'poptorch.MatMulSerializationMode'),
     ('py:class', 'poptorch.ReductionType'),
     ('py:class', 'poptorch.SyncPattern'),
+    ('py:class', 'poptorch.MeanReductionStrategy'),
     # Type hints
     ('py:data', 'typing.Optional'),
     ('py:data', 'typing.Callable'),